test: photo관련 test, service 테스트코드로 분리
