+++ src/main/java/dynamicquad/agilehub/issue/service/ImageService.java
@@ -19,13 +19,12 @@ public class ImageService {
     private final PhotoS3Manager photoS3Manager;
 
     public void saveImages(Issue issue, List<MultipartFile> files, String workingDirectory) {
-
+        log.info("save images");
         List<String> imagePath = photoS3Manager.uploadPhotos(files, workingDirectory);
-
+        log.info("save images = {} ", imagePath);
         if (imagePath.isEmpty()) {
             return;
         }
-        log.info("save images = {} ", imagePath);
 
         List<Image> images = imagePath.stream()
             .map(path -> Image.builder().path(path).build().setIssue(issue))
